Filtering

+++ ApiProject/Controllers/StockController.cs
@@ -1,5 +1,6 @@
 ﻿using ApiProject.Data;
 using ApiProject.DTOs.Stock;
+using ApiProject.Helpers;
 using ApiProject.Mappers;
 using ApiProject.Models;
 using ApiProject.Repositories.Interfaces;
@@ -12,18 +13,23 @@ namespace ApiProject.Controllers
     [ApiController]
     public class StockController : Controller
     {
-    
+
         private readonly IStockRepository _stockRepo;
-        public StockController(ApplicationDbContext context,IStockRepository stockRepo)
-        {            
+        public StockController(ApplicationDbContext context, IStockRepository stockRepo)
+        {
             _stockRepo = stockRepo;
         }
 
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
-            List<Stock> stocks = await _stockRepo.GetAllAsync();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+        
+            List<Stock> stocks = await _stockRepo.GetAllAsync(query);
 
             var result = stocks.Select(s => s.ToStockDto());
 
@@ -33,50 +39,66 @@ namespace ApiProject.Controllers
 
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             Stock stock = await _stockRepo.GetByIdAsync(id);
 
-            if(stock!=null)
+            if (stock != null)
             {
                 return Ok(stock.ToStockDto());
             }
             else
             {
-                return NotFound ();    
+                return NotFound();
             }
-             
+
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockDto stockDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             Stock stockModel = stockDto.ToStockFromCreateDto();
 
-            await _stockRepo.CreateAsync(stockModel);   
-        
+            await _stockRepo.CreateAsync(stockModel);
 
-            return CreatedAtAction(nameof(GetById),new { id = stockModel.Id },stockModel.ToStockDto());   
+
+            return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockDto updateDto)
         {
-            Stock stockModel = await _stockRepo.UpdateAsync(id,updateDto);  
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            Stock stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             if (stockModel == null)
             {
                 return NotFound();
-            }   
+            }
 
             return Ok(stockModel.ToStockDto());
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             Stock stockToDelete = await _stockRepo.DeleteAsync(id);
             if (stockToDelete == null)
             {
